+++ core/src/main/java/org/eqasim/core/analysis/trips/TripReaderFromEvents.java
@@ -2,6 +2,7 @@ package org.eqasim.core.analysis.trips;
 
 import java.util.Collection;
 
+import org.eqasim.core.components.transit.events.PublicTransitEvent;
 import org.matsim.core.api.experimental.events.EventsManager;
 import org.matsim.core.events.EventsUtils;
 import org.matsim.core.events.MatsimEventsReader;
@@ -16,7 +17,9 @@ public class TripReaderFromEvents {
 	public Collection<TripItem> readTrips(String eventsPath) {
 		EventsManager eventsManager = EventsUtils.createEventsManager();
 		eventsManager.addHandler(tripListener);
-		new MatsimEventsReader(eventsManager).readFile(eventsPath);
+		MatsimEventsReader matsimEventsReader = new MatsimEventsReader(eventsManager);
+		matsimEventsReader.addCustomEventMapper(PublicTransitEvent.TYPE, PublicTransitEvent::convert);
+		matsimEventsReader.readFile(eventsPath);
 		return tripListener.getTripItems();
 	}
 }