+++ CHANGELOG.md
@@ -5,7 +5,11 @@ This change log is written in descending order. Changes that happen after versio
 included in the (note yet determined) next version number.
 
 **Development version**
- 
+
+- Ensure outside activity id doesn't already exist
+- Network-based (car) routing now generates access and egress walk legs
+- Convert initial-routing only-walk legs to actual walk (instead of transit)
+- Don't put activities on motorway/trunk/link in the network
 - Updated to MATSim 14
 - Isolated the mode choice model in a standalone runnable script
 - Fixed LegIndex count between iterations in legs analysis